% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_AR_part.R
\name{apply_AR_part}
\alias{apply_AR_part}
\title{Apply the autoregressive recursion in the temporal domain}
\usage{
apply_AR_part(fts_x, operators_ar)
}
\arguments{
\item{fts_x}{The functional time series, represented as a matrix of size (\code{n_grid},\code{t_max}), which the autoregressive recursion is applied to.}

\item{operators_ar}{The list of length 'p' the order of the autoregressive part. The autoregressive operators are considered to be integral operators defined through their kernels which are saved as the elements of the list \code{operators_ar} as functions of two variables, \code{x} and \code{y}, returning the value of the kernel at point (\code{x},\code{y}). In case of degenerate autoregressive part define \code{operators_ar} as an empty list.}
}
\value{
functional time series sample, matrix of size (\code{n_grid},\code{t_max}) where these sample size parameters are copied from the input parameter \code{fts_x}
}
\description{
This function is not meant to be used by the end-user. It is internal function for the runtime of the hybrid simulation withing the FARFIMA_simulate function.
}
\references{
Rubin, Panaretos. \emph{Simulation of stationary functional time series with given spectral density}. arXiv, 2020
}
\seealso{
\code{\link{FARFIMA_simulate}}
}
