% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_eig.R
\name{BM_eig}
\alias{BM_eig}
\title{Brownian motion kernel eigendecomposition}
\usage{
BM_eig(n_grid)
}
\arguments{
\item{n_grid}{The spatial resolution of the discretization of [0,1], equidistant grid.}
}
\value{
Returns a list of two elements, \code{vectors} is a matrix of size (\code{n_grid},\code{n_grid}) containing the eigenfunctions in the columns of this matrix, and \code{values} vector of eigenvalues.
}
\description{
This function is not meant to be used by the end-user. It internally provides the decomposition of the Brownian motion for \code{\link{FARFIMA_simulate}} to be run with the Brownian motion covariance.
}
\references{
Rubin, Panaretos. \emph{Simulation of stationary functional time series with given spectral density}. arXiv, 2020
}
\seealso{
\code{\link{FARFIMA_simulate}}
}
