% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKL_covlagh_operator.R
\name{CKL_covlagh_operator}
\alias{CKL_covlagh_operator}
\title{Calculate the lag-h autocovariance operator for functional time series dynamics defined by its harmonic Karhunen-Loeve expansion.}
\usage{
CKL_covlagh_operator(
  harmonic_eigenvalues,
  harmonic_eigenfunctions,
  lag,
  n_grid,
  n_pc,
  n_grid_freq = 1000
)
}
\arguments{
\item{harmonic_eigenvalues}{function of two variables, \code{omega} and \code{n}, that assigns the \code{n}-th harmonic eigenvalue at frequency \code{omega}. Must be well defined for frequencies (0,pi]. The interval [pi,2pi) is not used and is calculated by mirroring of (0,pi]..}

\item{harmonic_eigenfunctions}{function of three variables, \code{omega}, \code{n} and \code{x}, that assigns the \code{n}-th harmonic eigenfunction at point \code{x} in [0,1] at frequency \code{omega}. Must be well defined for frequencies (0,pi]. The interval [pi,2pi) is not used and is calculated by mirroring of (0,pi].}

\item{lag}{The lag of the autocovariance to evaluate.}

\item{n_grid}{Number of grid points (spatial resolution) of the discretisation of [0,1]^2 for the operator kernel to evaluate.}

\item{n_pc}{The number of harmonic eigenfunctions to be used for the numerical integration at each frequency.}

\item{n_grid_freq}{The grid points for the spectral density to evaluate at. Partition of [0,pi].}
}
\value{
lag-h autocovariance operator, matrix of size (\code{n_grid},\code{n_grid})
}
\description{
Numerically calculate the lag-h covariance operators for functional time series dynamics defined by its harmonic Karhunen-Loeve expansion.
The calculation is done by numerically integrating the inverse formula, i.e. the spectral density multiplied by exp(-1i*lag*omega)
}
\examples{
# Define the eigenvalues and eigenfunction of the functional time series
harmonic_eigenvalues <- function( omega, n ){ 1/( (1-0.9 *cos(omega)) * (n*pi)^2 ) }
harmonic_eigenfunctions <- function(omega, n, x){ sqrt(2)*sin( n*(pi*x-omega)  ) }

# evaluation setting
lag <- 1 # change here to evaluate different lag-h autocovariance operator. put "lag <- 0" for lag-0 covariance operator
n_grid <- 101
n_pc <- 100

# calculate the lag-h autocovariance operator
covlagh <- HKL_covlagh_operator(harmonic_eigenvalues, harmonic_eigenfunctions, lag, n_grid, n_pc)

# visualise as a surface plot
persp(covlagh)

}
\references{
Rubin, Panaretos. \emph{Simulation of stationary functional time series with given spectral density}. arXiv, 2020
}
\seealso{
\code{\link{HKL_simulate}}
}
