% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_density_covlagh_operator.R
\name{spec_density_covlagh_operator}
\alias{spec_density_covlagh_operator}
\title{Calculate the lag-h autocovariance operator for FTS defined by its spectral density operator}
\usage{
spec_density_covlagh_operator(spec_density, lag, n_grid, n_grid_freq = 2000)
}
\arguments{
\item{spec_density}{The spectral density operator defined as an integral operator through with the given kernel. Function of three variables, \code{omega}, \code{x}, \code{y}, that assigns the value of the spectral density kernel at point (\code{x},\code{y}) in [0,1]^2 at frequency \code{omega} in [0,pi]. Must be well defined for frequencies (0,pi]. The interval [pi,2pi) is not used and is calculated by mirroring of (0,pi].}

\item{lag}{The lag of the autocovariance to evaluate.}

\item{n_grid}{Number of grid points (spatial resolution) of the discretisation of [0,1]^2 for the operator kernel to evaluate.}

\item{n_grid_freq}{The grid points for the spectral density to evaluate at. Partition of [0,pi].}
}
\value{
lag-h autocovariance operator, matrix of size (\code{n_grid},\code{n_grid})
}
\description{
Numerically calculate the lag-h covariance operators for functional time series dynamics defined directly by its spectral density operator.
The calculation is done by numerically integrating the inverse formula, i.e. the spectral density multiplied by exp(-1i*lag*omega)
}
\examples{
# Define the spectral density operator as an integral operator with kernel
k_bbridge <- function(x,y) { pmin(x,y)-x*y }
spec_density <- function( omega, x,y ){ 1/(1-0.9 *cos(omega)) * k_bbridge( (x-omega/pi)\%\%1, (y-omega/pi)\%\%1  ) }

# evaluation setting
lag <- 1 # change here to evaluate different lag-h autocovariance operator. put "lag <- 0" for lag-0 covariance operator
n_grid <- 101

# calculate the lag-h autocovariance operator
covlagh <- spec_density_covlagh_operator(spec_density, lag, n_grid)

# visualise as a surface plot
persp(covlagh)

}
\references{
Rubin, Panaretos. \emph{Simulation of stationary functional time series with given spectral density}. arXiv, 2020
}
\seealso{
\code{\link{CKL_simulate}}
}
